Add treesitter based render support for a document
Introduce a second render "engine" for handling better the HTML based
hoogle documents. This render uses neovim's builtin treesitter
functionality to parse the hoogle document and uses the AST to do the
highlighting.

This feature requires that treesitter's html language parser is installed,
by default that is not a case. However, that can be done easily with
nvim-treesitter plugin.

Additionally, this include a feature to remove the hoogle's own text
wrapping, and that can be left to do for telescope/neovim.